adapting controller and entities

+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/service/RatingService.java
@@ -11,7 +11,7 @@ import javax.transaction.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class RatingService extends CrudService<Rating, Integer> {
+public class RatingService extends CrudService<Rating, Long> {
     @Getter
     private final RatingRepository repository;
 }
